requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.34
psycopg2-binary==2.9.9
pydantic==2.9.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
redis==5.0.8
rq==1.16.2
alembic==1.13.2
httpx==0.27.2
python-dotenv==1.0.1







